+++ MORE/Events/CombatDataCalibrationEvent.cs
@@ -16,6 +16,7 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 	public static void Register(IPluginPackage<IModManifest> package, IModHelper helper)
 	{
 		CombatAnalyzerGadgetArtifact.RegisterArtifact(helper);
+		WeakCombatAnalyzerGadgetArtifact.RegisterArtifact(helper);
 
 		EventName = $"{package.Manifest.UniqueName}::{MethodBase.GetCurrentMethod()!.DeclaringType!.Name}";
 
@@ -56,10 +57,30 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 			],
 			choiceFunc = EventName
 		};
+		DB.story.all[$"{EventName}::Concerns"] = new()
+		{
+			type = NodeType.@event,
+			lines = [
+				new CustomSay
+				{
+					who = "scientist",
+					loopTag = "neutral",
+					flipped = true,
+					Text = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Concerns-1-Bjorn"]),
+				},
+				new CustomSay
+				{
+					who = "scientist",
+					loopTag = "neutral",
+					flipped = true,
+					Text = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Concerns-2-Bjorn"]),
+				},
+			],
+			choiceFunc = $"{EventName}::Concerns"
+		};
 		DB.story.all[$"{EventName}::Yes"] = new()
 		{
 			type = NodeType.@event,
-			oncePerRun = true,
 			lines = [
 				new CustomSay
 				{
@@ -73,7 +94,6 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 		DB.story.all[$"{EventName}::No"] = new()
 		{
 			type = NodeType.@event,
-			oncePerRun = true,
 			lines = [
 				new CustomSay
 				{
@@ -86,6 +106,7 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 		};
 
 		DB.eventChoiceFns[EventName] = AccessTools.DeclaredMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(GetChoices));
+		DB.eventChoiceFns[$"{EventName}::Concerns"] = AccessTools.DeclaredMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(GetConcernsChoices));
 	}
 
 	public static void UpdateSettings(IPluginPackage<IModManifest> package, IModHelper helper, ProfileSettings settings)
@@ -102,7 +123,7 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 	[UsedImplicitly]
 	private static List<Choice> GetChoices(State state)
 		=> [
-			new Choice
+			new()
 			{
 				label = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Choice-Yes"]),
 				key = $"{EventName}::Yes",
@@ -110,11 +131,42 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 					new AAddArtifact { artifact = new CombatAnalyzerGadgetArtifact() }
 				]
 			},
-			new Choice
+			new()
+			{
+				label = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Choice-Concerns"]),
+				key = $"{EventName}::Concerns"
+			},
+			new()
 			{
 				label = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Choice-No"]),
 				key = $"{EventName}::No"
-			}
+			},
+		];
+
+	[UsedImplicitly]
+	private static List<Choice> GetConcernsChoices(State state)
+		=> [
+			new()
+			{
+				label = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Concerns-Choice-Yes"]),
+				key = $"{EventName}::Yes",
+				actions = [
+					new AAddArtifact { artifact = new WeakCombatAnalyzerGadgetArtifact() }
+				]
+			},
+			new()
+			{
+				label = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Concerns-Choice-Science"]),
+				key = $"{EventName}::Yes",
+				actions = [
+					new AAddArtifact { artifact = new CombatAnalyzerGadgetArtifact() }
+				]
+			},
+			new()
+			{
+				label = ModEntry.Instance.Localizations.Localize(["event", "CombatDataCalibration", "Concerns-Choice-No"]),
+				key = $"{EventName}::No"
+			},
 		];
 
 	private sealed class CombatAnalyzerGadgetArtifact : Artifact
@@ -135,6 +187,9 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 				Description = ModEntry.Instance.AnyLocalizations.Bind(["event", "CombatDataCalibration", "artifact", "description"]).Localize
 			});
 		}
+
+		public override List<Tooltip> GetExtraTooltips()
+			=> StatusMeta.GetTooltips(Status.powerdrive, 1);
 
 		public override void OnCombatStart(State state, Combat combat)
 		{
@@ -159,6 +214,58 @@ internal sealed class CombatDataCalibrationEvent : IRegisterable
 			{
 				amount = 3,
 				limitPools = [ArtifactPool.Boss]
+			});
+		}
+	}
+	
+	private sealed class WeakCombatAnalyzerGadgetArtifact : Artifact
+	{
+		public static void RegisterArtifact(IModHelper helper)
+		{
+			ArtifactEntry = helper.Content.Artifacts.RegisterArtifact(MethodBase.GetCurrentMethod()!.DeclaringType!.Name, new()
+			{
+				ArtifactType = MethodBase.GetCurrentMethod()!.DeclaringType!,
+				Meta = new()
+				{
+					owner = Deck.colorless,
+					pools = [ArtifactPool.EventOnly],
+					unremovable = true
+				},
+				Sprite = helper.Content.Sprites.RegisterSprite(ModEntry.Instance.Package.PackageRoot.GetRelativeFile("assets/Artifact/WeakCombatAnalyzerGadget.png")).Sprite,
+				Name = ModEntry.Instance.AnyLocalizations.Bind(["event", "CombatDataCalibration", "artifactWeak", "name"]).Localize,
+				Description = ModEntry.Instance.AnyLocalizations.Bind(["event", "CombatDataCalibration", "artifactWeak", "description"]).Localize
+			});
+		}
+
+		public override List<Tooltip> GetExtraTooltips()
+			=> StatusMeta.GetTooltips(Status.overdrive, 1);
+
+		public override void OnTurnStart(State state, Combat combat)
+		{
+			base.OnTurnStart(state, combat);
+			if (combat.turn != 1)
+				return;
+			
+			combat.Queue(new AStatus
+			{
+				targetPlayer = false,
+				status = Status.overdrive,
+				statusAmount = 1,
+				artifactPulse = Key()
+			});
+		}
+
+		public override void OnCombatEnd(State state)
+		{
+			base.OnCombatEnd(state);
+			if (state.map.markers[state.map.currentLocation].contents is not MapBattle { battleType: BattleType.Boss })
+				return;
+
+			state.rewardsQueue.Queue(new ALoseArtifact { artifactType = Key() });
+			state.rewardsQueue.Queue(new AArtifactOffering
+			{
+				amount = 3,
+				limitPools = [ArtifactPool.Common]
 			});
 		}
 	}